restructure datastore utils code and tests
